Release 1 | Beta 1 | Development 3 (V0.3.3)
* Added FramebufferStoreCubemap
* Added Text
* Added point light shadows
* Added methods to Font that only update either the position
  that text will be rendered or the actual text to be rendered
* Added 'getCurrentFrameInAnimation' method in Animation2D
* Added ability to add fonts to ResourcePack
* Added 1440p to VideoResolution

* Changed loadShader method in Shader to load a geometry
  shader if it is also availabl

* Renamed 'timeBetweenFrame' to 'timeBetweenFrames' in
  Animation2D

* Cleaned up RenderScene3D slightly

* Increased maximum number of bones in a model from 80 to 90

* Fixed bug: RenderScene3D didn't use wireframes correctly
             when not using lighting
* Fixed bug: Rotations for skeletal animation weren't
             interpolated correctly
* Fixed bug: Depth sorting for particles ignored the last
  			 particle
* Fixed bug: Font getHeight only measured the height of one
             line if \n was included

* Note: Text handles rendering of a Font instead of Font

+++ UnnamedEngine/Source/core/Vector.h
@@ -156,7 +156,7 @@ public:
 		float theta = acosf(dot) * factor;
 		Vector<T, N> relative = b - a * dot;
 		relative.normalise();
-		return ((a * cosf(theta)) + (relative * sin(theta)));
+		return ((a * cosf(theta)) + (relative * sinf(theta)));
 	}
 	inline static Vector<T, N> nlerp(const Vector<T, N>& a, const Vector<T, N>& b, T factor) { return lerp(a, b, factor).normalise(); }
